Updated the documentation to v0.9.0.
This was not a trivial exercise. The docs don't build without adding an
import "numpydoc" in the sphinx config. This results in ugly html
"method" names all over the place, which have been removed with:

    sed -i '/<p class="rubric">Methods<\/p>/d' */*.html

Some of the examples don't run correctly with matplotlib v1.3, others
don't run without matplotlib v1.3.

Many of the images that are generated are still unused and I've manually
removed those from this commit. That has been a process which involves
variants of the following command:

    ls -1 examples/*.p?? | grep -v thumb | xargs rm

That is pretty much all. The doc building process in the future will be
improved such that it builds seamlessly with matplotlib v1.3+